Use Fixtures instead
