+++ tests/Integration/Http/Fixtures/EmptyPostCollectionResource.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Illuminate\Tests\Integration\Http\Fixtures;
+
+use Illuminate\Http\Resources\Json\ResourceCollection;
+
+class EmptyPostCollectionResource extends ResourceCollection
+{
+    public $collects = PostResource::class;
+}
