+++ tests/Integration/Http/Fixtures/JsonSerializableResource.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace Illuminate\Tests\Integration\Http\Fixtures;
+
+use JsonSerializable;
+
+class JsonSerializableResource implements JsonSerializable
+{
+    public $resource;
+
+    public function __construct($resource)
+    {
+        $this->resource = $resource;
+    }
+
+    public function jsonSerialize()
+    {
+        return [
+            'id' => $this->resource->id,
+        ];
+    }
+}
