+++ tests/Integration/Http/Fixtures/PostCollectionResource.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Illuminate\Tests\Integration\Http\Fixtures;
+
+use Illuminate\Http\Resources\Json\ResourceCollection;
+
+class PostCollectionResource extends ResourceCollection
+{
+    public $collects = PostResource::class;
+
+    public function toArray($request)
+    {
+        return ['data' => $this->collection];
+    }
+}
