+++ tests/Integration/Http/Fixtures/PostResourceWithExtraData.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Illuminate\Tests\Integration\Http\Fixtures;
+
+class PostResourceWithExtraData extends PostResource
+{
+    public function with($request)
+    {
+        return ['foo' => 'bar'];
+    }
+}
