+++ tests/Integration/Http/Fixtures/PostResourceWithOptionalRelationship.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Illuminate\Tests\Integration\Http\Fixtures;
+
+class PostResourceWithOptionalRelationship extends PostResource
+{
+    public function toArray($request)
+    {
+        return [
+            'id' => $this->id,
+            'comments' => new CommentCollection($this->whenLoaded('comments')),
+        ];
+    }
+}
