+++ tests/Integration/Http/Fixtures/ReallyEmptyPostResource.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Illuminate\Tests\Integration\Http\Fixtures;
+
+use Illuminate\Http\Resources\Json\Resource;
+
+class ReallyEmptyPostResource extends Resource
+{
+    //
+}
